Moved generateAuthUri and generateAuthCodeVerifier to NsoConnector class
