+++ libs/splatnet/src/index.ts
@@ -2,3 +2,4 @@ export * from './lib/connect/generate-iksm.module';
 export * from './lib/connect/generate-iksm.model';
 export * from './lib/connect/NsoConnector.class';
 export * from './lib/NsoError';
+export * from './lib/connect/nso-app-version'
